client: add new routine to get fscid from a ceph_mount
Ceph recently gained the ability to create alternate named filesystems
within a cluster. We want to allow nfs-ganesha to export them, but it
currently generates filehandles using a inode number+snapid tuple, and
that will not be unique across multiple filesystems.

Add a new field to hold the fscid in the Client, as that is generated
on a per-filesystem basis via monotonic counter. When we mount, fetch
the fscid and store it in that field. Add a new Client accessor, and a
libcephfs function that returns that value.

Tracker: http://tracker.ceph.com/issues/36585
Signed-off-by: Jeff Layton <jlayton@redhat.com>