mgr/dashboard: Reduce Jest logs in CI
With the current number of unit tests we have for the frontend the output from
Jest is too big, and is exceeding the limit from jenkins.
This is resulting in the omission of linting results.

With this change, Jest will only output the results of failing tests and will no
longer show the coverage in the logs.
Jenkins will keep tracking the coverage via the generated cobertura file.

Signed-off-by: Tiago Melo <tmelo@suse.com>