bluestore/NVMEDevice.cc: fix ceph_assert() when enable SPDK with 64KB kernel page size
When enable SPDK in Ceph with 64KB kernel page size, observed the
ceph_assert() in NVMEDevice. In SPDK NVME driver, the block size
should be the sector size of the NVMe device, not system page size.
Get the correct sector size by SPDK API.

This patch corrects the NVME block size and fixes the problem.

Fixes: http://tracker.ceph.com/issues/36624

Signed-off-by: tone.zhang <tone.zhang@arm.com>