tests: do not check for invalid k/m combinations
This reverts commit 4d7653633610ee6ef0f2c0ac5f12b134ebaf58a3.

in 6067f847 and 61bef0d8, we detect invalid "k". and we should always
detect invalid combinations of parameters in
ErasureCodeInterface::init(), and return -EINVAL instead of "autotune" them
as put in the commit message of 4d765363, if the profile is invalid for
given erasure plugin. so ErasureCodeBench::encode() and
ErasureCodeBench::decode() will bail out with the error message returned by
ErasureCodeInterface::init() if the parameters are not valid.

also, i checked all ErasureCodeInterface::init() implementations, none
of them is overwriting the settings without returning a non-zero return
code. so we are not overriding or autotuning the settings specified by
user.

Fixes: http://tracker.ceph.com/issues/16500
Signed-off-by: Kefu Chai <kchai@redhat.com>