rgw: archive: rework bucket removal
Keep reference to the original bucket in the new bucket instance object,
so that if a new bucket is created with the same name, we can generate
a name off the original name. Things could be much simpler if we could
just point a new bucket entrypoint at the old bucket entrypoint, but
we can't have (at the moment) bucket entrypoint that points at a bucket
instance with a different bucket name.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>