rgw: period update: check for dangling master zone references
If we are deleting a master zone of a zonegroup fail on period update, if this
was intentional, either creating / modifying a zone as master or in case of
deletions, deletion of the zonegroup itself will correct the period update to
work correctly. Without the check, while period commit will be successful a
subsequent RGWRados::init_complete() will fail.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>