mgr: Ignore daemon if no metadata was returned
It can happen that the Mgr does not return any metadata for a given
daemon as it might not be available at that moment.

None is returned by the get_metadata() method at that moment and both
the Influx and Telegraf module should then ignore the daemon in their
statistics and continue on to the next daemon.

Signed-off-by: Wido den Hollander <wido@42on.com>

(manual cherry-pick of 02569c88341a6dcc8aee5626f10ac2927c2cf064)

Signed-off-by: Nathan Cutler <ncutler@suse.com>