qa/deepsea: separate ses_qa.py task for non-DS testing
Many of the test cases we want to implement have nothing to do
with DeepSea except that DS is used to deploy the Ceph cluster.

Implement a separate task for these test cases, and reserve
deepsea.py as much as possible for DS-specific code.

Move the MGR plugins test cases into this new task (this includes
implementing a new test case for the influx plugin).

Signed-off-by: Nathan Cutler <ncutler@suse.com>