mgr/dashboard: Settings service
The settings service handles settings that are set
within the dashboard.

The service has a helpful function and a good memory. The function will
allow to check if a certain option is set, if it's set the callback will
be triggered. If the API request was made before for this particular
option, no API request is made anymore, as the service will remember if
the state is set or not.

Fixes: https://tracker.ceph.com/issues/37471
Signed-off-by: Stephan Müller <smueller@suse.com>