denc: enable decode basic_string<>
we have

buffer::list::iterator_impl<is_const>::copy(unsigned len,
std::string &dest)

but we don't have the templatized version of it. so, if we
plugin a different allocator to basic_string<>, the new string won't
decode with buffer::list::const_iterator. this decode variant is used if
the caller only has a (probably non-contiguous) buffer::list in hand.

in this change, copy(unsigned, char*) is used as an alternative.

Signed-off-by: Kefu Chai <kchai@redhat.com>