rgw: fix bad versioned bucket stats after reshard
When a versioned bucket is resharded, the stats for bucket index
entries of type PlainIdx and InstanceIdx were both accumulated to
determine the bucket stats. This caused a doubling of some stats, such
as bytes used. This fix makes certain that only PlainIdx entries are
accumulated. This works for unversioned buckets as well, as they only
have Plain Idx entries.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>