build, tests: introduce compiletest_cxx11_client for C++11 conformity.
This patch extends the usual pull request workflow with extra stage:
validation of the C++11 compliance of the public headers (currently
`buffers.h` only; I would love to have this broader) on `make check`.
The idea is to let the bots to detect breaking the language part of
our public API early, and thus off-load reviewers and Teuthology.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>