mgr/deepsea: check for inflight completions when starting event reader
This handles the case where an operation is inflight, but salt-api dies
(or the connection to salt-api has failed for some reason).  The next time
we're able to connect, we first check if there's any outstanding completions,
and look up matching salt jobs to see if they finished in the meantime.

Signed-off-by: Tim Serong <tserong@suse.com>