ceph-volume: batch mixed type scenarios have no need to calulate data size
We know with a mixed type scenario the device used for data will be used
at 100% capacity. This means we do not need to be explict when asking
for the size of the data lvs, which avoids rounding errors with very
small device sizes.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>
(cherry picked from commit 0023961397d19659200eaf420f2789db17a71b34)