ceph-volume: use Device.lvm_size in batch strategies
We should show the user what the size of the device will be after lvm
creates a pv out of it. This way there isn't a discrepency between the
sizes that are reported to the user and what is actually created.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>
(cherry picked from commit 071e7ce41a58702c0eb84c35bd9388e2ec24ac13)