ceph-volume: py3 - ensure correct string encoding
When writing data to stdout, ensure that strings are correctly
encoded for the encoding of the stdout;  systemd is good at
generating unicode characters which get captured by the process
module and then re-logged via terminal.

If the LANG environment variable is unset in the execution
environment this results in an encoding error during the write
operation as the unicode is not compatible with the 'ascii'
encoding of stdout in this situation.

Converting the string to the target encoding using 'replace'
rather than 'strict' makes sure the content can be written.

Signed-off-by: James Page <james.page@ubuntu.com>