msg/async/rdma: switch QP to error state prior actual destroy
Before destroy QP has to be switched to error state to guarantee all
outstanding tx WCEs are returned and SRQ is fully consumed. In this
patch a beacon is posted to the send queue just after QP has been
switched to error state.  Completion of a posted beacon indicates
queues are drained and QP can be safely destroyed.

This patch also removes handling of async events and accounting of tx
WRs, just because handled beacon acts as a last event when QP can
be destroyed.  The other reason why async events is not reliable is
that not all uverbs providers support them (e.g. qlogic, qedr).

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>