mon/MonCommand: populate 'availability' field for compat
The CLI will behave if this field is not included in teh JSON, but when you have
a nautilus+ leader mon and a luminous or mimic peon mon they will decode this empty
field and include "avail": "" in the JSON command descriptions, which the CLI does
not like.

Fix by encoding "cli,rest" in this field for the benefit of old mons.

Fixes: http://tracker.ceph.com/issues/37583
Signed-off-by: Sage Weil <sage@redhat.com>