rgw: fix versioned obj copy generating tags
We fail to generate obj_tag and tail tags for a versioned object copy, as we
don't set the object instance until later and the atomic state hence isn't set
for this object as we call `prepare_atomic_modification` which checks for
atomicity before setting write tags. Since the tag is later used by the GC
processor for refcounting, versioned objects are not cleaned up leaving orphaned
tail instances behind. For older objects these will have to be manually cleaned
up, while this commit should fix the issues for newer versioned copy

Fixes: http://tracker.ceph.com/issues/37588
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>