pybind/mgr/mgr_module: normalize defaults to str when type is missing
If the module doesn't specify a type for the option, it is probably
an older module.  Make sure that the default value we provide is normalized
to a string, just in case the module declares something like

  {
     'name': 'my_option',
     'default': True,
  }

where the default is *not* a string, but an user-set value coming from
the config infrastructure would be a string.

Signed-off-by: Sage Weil <sage@redhat.com>